Feature(MasterModule): add untested sync. of Dude params
In function CTOSMasterClient::PrepareDude(), the Dude is prepared before transferring control of the slave entity to him. During preparation, its properties are saved on the server and after Dude has stopped managing the slave, Dude also needs to be prepared, but his old state must be returned. The saved settings are now applied to Dude.

Applicable parameters:
- position
- rotation
- species
- suit mode
- suit energy